Refactor main.py into 2 files
Refactor `main.py` by moving UI-related functions to a new file `app_ui.py`.

* **Add `app_ui.py`**:
  - Add `create_video_card`, `show_credentials_instructions`, `show_login_ui`, `show_main_ui`, `parse_duration`, and `relative_time` functions.

* **Modify `main.py`**:
  - Remove `create_video_card`, `show_credentials_instructions`, `show_login_ui`, and `show_main_ui` functions.
  - Import `create_video_card`, `show_credentials_instructions`, `show_login_ui`, and `show_main_ui` from `app_ui.py`.
  - Update `main` function to call `show_main_ui` and `show_login_ui` with `app` as an argument.

---

For more details, open the [Copilot Workspace session](https://copilot-workspace.githubnext.com/xbito/yt_gt_browser?shareId=XXXX-XXXX-XXXX-XXXX).
